requirements.txt
Django==4.2
djangorestframework==3.15.2
psycopg2==2.9.6
django-environ==0.10.0
elasticsearch-dsl==8.14.0
celery==5.4.0
boto3==1.34.145
django-celery-results==2.5.1